# ⚡ Better Faster Cleaner `STATUS` parsing