Although "number" is still the default `status-att-val`, this uses
ExtensionData with RFC4466's `tagged_ext_val` for any unknown
non-numeric `STATUS` attribute.

Running the benchmarks (on my phone, without YJIT) shows a 40% speedup!

    invalid_status_response_trailing_space
          v0.4.4-16-g0be6b65b:     43956.1 i/s
                        0.4.4:     31788.6 i/s - 1.38x  slower

     rfc3501_7.2.4_STATUS_response_example
          v0.4.4-16-g0be6b65b:     45436.2 i/s
                        0.4.4:     32458.5 i/s - 1.40x  slower

       status_response_uidnext_uidvalidity
          v0.4.4-16-g0be6b65b:     45334.2 i/s
                        0.4.4:     32709.1 i/s - 1.39x  slower